+++ desktop/php/events.php
@@ -12,19 +12,54 @@ if (!isConnect('admin')) {
 
 <div class="col-lg-12"><br>
 	<br>
-	<div class="input-group" style="display:inline-flex">
+	<div class="input-group" style="margin-bottom:20px">
 		<span class="input-group-btn">
 			<a class="btn geoAction rounded" id="gotoHome"><i class="fa fa-arrow-circle-left"></i> retour </a>
 		</span>
 	</div>
-	<br>
-	<br>
 	<?php
 	$events = frigate::showEvents();
+
+	echo '<div class="col-sm-9" style="margin-bottom:5px">';
+  	echo '<a class="btn btn-info button-xs" id="selectAllCameras" style="margin-right:10px"><i class="fas fa-check"></i> {{Tout}}</a>';
+    echo '<a class="btn btn-info button-xs" id="deselectAllCameras" style="margin-right:20px"><i class="fas fa-times"></i> {{Aucun}}</a>';
+	$cameras = array_unique(array_column($events, 'camera'));
+    foreach ($cameras as $camera) {
+        echo '<label class="checkbox-inline"><input type="checkbox" class="eqLogicAttr cameraFilter" value="' . $camera . '" checked> ' . ucfirst($camera) . '</label>';
+    }
+	echo '</div>';
+
+	echo '<div class="col-sm-9" style="margin-bottom:20px">';
+  	echo '<a class="btn btn-info button-xs" id="selectAllLabels" style="margin-right:10px"><i class="fas fa-check"></i> {{Tout}}</a>';
+    echo '<a class="btn btn-info button-xs" id="deselectAllLabels" style="margin-right:20px"><i class="fas fa-times"></i> {{Aucun}}</a>';
+	$labels = array_unique(array_column($events, 'label'));
+    foreach ($labels as $label) {
+        echo '<label class="checkbox-inline"><input type="checkbox" class="eqLogicAttr labelFilter" value="' . $label . '" checked> ' . ucfirst($label) . '</label>';
+    }
+	echo '</div>';
+
+	echo '<div class="col-sm-9" style="margin-bottom:5px">
+        <label>Entre <input type="datetime-local" id="startDate"></label>
+        <label>et <input type="datetime-local" id="endDate"></label>
+    </div>
+
+    <div class="col-sm-10" style="margin-bottom:5px">
+        Date de 
+        <label><input type="radio" name="timeFilter" value="" checked> Tous</label>
+        <label><input type="radio" name="timeFilter" value="1h"> Moins d\'une heure</label>
+        <label><input type="radio" name="timeFilter" value="2h"> Moins de deux heures</label>
+        <label><input type="radio" name="timeFilter" value="6h"> Moins de six heures</label>
+        <label><input type="radio" name="timeFilter" value="12h"> Moins de douze heures</label>
+        <label><input type="radio" name="timeFilter" value="1d"> Moins d\'un jour</label>
+        <label><input type="radio" name="timeFilter" value="2d"> Moins de deux jours</label>
+        <label><input type="radio" name="timeFilter" value="1w"> Moins d\'une semaine</label>
+    </div>';
+    
+	echo '<div>';
 	foreach ($events as $event) {
 		//div globale start
-		echo '<div class="col-lg-4 ">';
-		echo '<div class="col-lg-12 frigateEvent">';
+		echo '<div data-date="' . $event['date'] .  '" data-camera="' . $event['camera'] .  '" data-label="' . $event['label'] .  '" class="frigateEventContainer col-lg-4 ">';
+      	echo '<div class="col-lg-12 frigateEvent">';
 		// div img
 		echo '<div>';
 		echo '<img class="imgSnap" src="' . $event['img'] . '"/>';
@@ -37,18 +72,21 @@ if (!isConnect('admin')) {
 		echo '<i class="fas fa-clock"></i><span>  ' . $event['date'] . ' (' . $event['duree'] . ' sc)</span>';
 		echo '</div>';
 		// div buttons
-		echo '<div class="eventBtn">';
+		echo '<div class="eventBtns"';
+        if ($event['hasSnapshot'] == 1) echo ' data-snapshot="' . $event['snapshot'] . '"';
+        if ($event['hasClip'] == 1) echo ' data-video="' . $event['clip'] . '"';
+        echo '>';
 		if ($event['hasSnapshot'] == 1) {
-			echo '<button class="hover-button" onclick="openClip(this)" id="' . $event['snapshot'] . '">';
+			echo '<button class="hover-button snapshot-btn">';
 			echo '<i class="fas fa-image"></i>';
 			echo '</button>';
 		}
 		if ($event['hasClip'] == 1) {
-			echo '<button class="hover-button" onclick="openSnapshot(this)" id="' . $event['clip'] . '">';
+			echo '<button class="hover-button video-btn">';
 			echo '<i class="fas fa-camera"></i>';
 			echo '</button>';
 		}
-		echo '<button class="hover-button" onclick="deleteEvent(this)" id="' . $event['id'] . '" title="Supprimer l\'event sur votre serveur frigate">';
+		echo '<button class="hover-button" onclick="deleteEvent(\'' . $event['id'] . '\')" title="Supprimer l\'event sur votre serveur frigate">';
 		echo '<i class="fas fa-trash"></i>';
 		echo '</button>';
 		echo '</div>';
@@ -56,7 +94,26 @@ if (!isConnect('admin')) {
 		echo '</div>';
 		echo '</div>';
 	}
+	echo '</div>';
+
 	?>
+
+    <div id="mediaModal" class="modal">
+        <div class="modal-content">
+            <span class="close">&times;</span>
+            <button id="showVideo" class="hidden-btn">Voir la vidéo</button>
+            <button id="showImage" class="hidden-btn">Voir la snapshot</button>
+            <div class="video-container active">
+                <video id="videoPlayer" width="100%" controls autoplay>
+                    <source id="videoSource" src="" type="video/mp4">
+                    Votre navigateur ne supporte pas la balise vidéo.
+                </video>
+            </div>
+            <div class="image-container">
+                <img id="snapshotImage" src="" alt="Snapshot" width="100%">
+            </div>
+        </div>
+    </div>
 
 </div>
 
@@ -70,9 +127,9 @@ if (!isConnect('admin')) {
 	}
 
 	.imgSnap {
-		display: flex: 0 0 auto;
+		flex: 0 0 auto;
 		position: relative;
-		//   background-color: rgb(var(--defaultBkg-color));
+		background-color: rgb(var(--defaultBkg-color));
 		margin-left: -15px;
 		height: 125px;
 		border-bottom-left-radius: 10px;
@@ -81,13 +138,13 @@ if (!isConnect('admin')) {
 	}
 
 	.eventText {
-		display: flex: 1 1 auto;
+		flex: 1 1 auto;
 		position: relative;
 		margin-left: 20px;
 
 	}
 
-	.eventBtn {
+	.eventBtns {
 		display: flex;
 		flex-direction: column;
 		align-items: center;
@@ -113,6 +170,51 @@ if (!isConnect('admin')) {
 		align-items: flex-end;
 		/* Align buttons to the right */
 	}
+
+    .eventHidden {
+      display: none;
+    }
+
+    .modal {
+      display: none;
+      position: fixed;
+      z-index: 2;
+      left: 0;
+      top: 0;
+      width: 100%;
+      height: 100%;
+      overflow: auto;
+      background-color: rgba(0,0,0,0.7);
+      padding-top: 60px;
+    }
+    .modal-content {
+      background-color: #fefefe;
+      margin: 5% auto;
+      padding: 20px;
+      border: 1px solid #888;
+        width: 80%;
+    }
+    .close {
+      color: #aaa;
+      float: right;
+      font-size: 28px;
+      font-weight: bold;
+    }
+    .close:hover,
+    .close:focus {
+      color: black;
+      text-decoration: none;
+      cursor: pointer;
+    }
+    .video-container, .image-container {
+      display: none;
+    }
+    .active {
+      display: block;
+    }
+    .hidden-btn {
+      display: none;
+    }
 </style>
 
 <?php include_file('desktop', 'events', 'js', 'frigate'); ?>