+++ docs/fr_FR/changelog.md
@@ -5,5 +5,12 @@
 >S'il n'y a pas d'information sur la mise à jour, c'est que celle-ci concerne uniquement de la mise à jour de documentation, de traduction ou de texte.
 
 
+# 15/06/2024 Beta 0..0
+- Ajout de logs
+- Ajout dans l'équipement Event de commande pour activer le cron
+- Modification de la configuration cron, utiliser les checkbox jeedom.
+- Ajout d'options page events (meri @noodom)
+- Configuration pièce par default
+
 # 15/06/2024 Beta 0.5.0
 - première version Beta